@@ -1,27 +1,27 @@
 module github.com/guacsec/guac
 
-go 1.18
+go 1.19
 
 require (
-	cloud.google.com/go/storage v1.28.1
-	github.com/fsouza/fake-gcs-server v1.43.0
+	cloud.google.com/go/storage v1.29.0
+	github.com/fsouza/fake-gcs-server v1.44.0
 	github.com/in-toto/in-toto-golang v0.3.4-0.20220709202702-fa494aaa0add
 	github.com/neo4j/neo4j-go-driver/v4 v4.4.4
 	github.com/secure-systems-lab/go-securesystemslib v0.4.0
 	github.com/spf13/cobra v1.6.1
 	go.uber.org/zap v1.24.0
-	google.golang.org/api v0.106.0
+	google.golang.org/api v0.108.0
 )
 
 require (
-	cloud.google.com/go v0.105.0 // indirect
+	cloud.google.com/go v0.107.0 // indirect
 	cloud.google.com/go/compute v1.14.0 // indirect
 	cloud.google.com/go/iam v0.8.0 // indirect
 	cloud.google.com/go/pubsub v1.28.0 // indirect
 	github.com/felixge/httpsnoop v1.0.3 // indirect
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/google/go-cmp v0.5.9
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/googleapis/enterprise-certificate-proxy v0.2.1 // indirect
 	github.com/googleapis/gax-go/v2 v2.7.0 // indirect
