@@ -161,41 +161,22 @@ func Test_Publish(t *testing.T) {
 	ctx, cancel = context.WithTimeout(ctx, time.Second)
 	defer cancel()
 
-	docChan := make(chan processor.DocumentTree, 1)
-	errChan := make(chan error, 1)
-	defer close(errChan)
-	defer close(docChan)
-
-	go func() {
-		errChan <- testSubscribe(ctx, docChan)
-	}()
-
-	numSubscribers := 1
-	subscribersDone := 0
-
-	for subscribersDone < numSubscribers {
-		select {
-		case d := <-docChan:
-			if !dochelper.DocTreeEqual(d, expectedDocTree) {
-				t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", dochelper.StringTree(d), dochelper.StringTree(expectedDocTree))
-			}
-		case err := <-errChan:
-			if err != nil && !errors.Is(err, context.DeadlineExceeded) {
-				t.Errorf("nats emitter Subscribe test errored = %v", err)
-			}
-			subscribersDone += 1
-		}
-	}
-	for len(docChan) > 0 {
-		d := <-docChan
+	transportFunc := func(d processor.DocumentTree) error {
 		if !dochelper.DocTreeEqual(d, expectedDocTree) {
 			t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", dochelper.StringTree(d), dochelper.StringTree(expectedDocTree))
 		}
+		return nil
 	}
 
+	err = testSubscribe(ctx, transportFunc)
+	if err != nil {
+		if err != nil && !errors.Is(err, context.DeadlineExceeded) {
+			t.Errorf("nats emitter Subscribe test errored = %v", err)
+		}
+	}
 }
 
-func testSubscribe(ctx context.Context, docChannel chan<- processor.DocumentTree) error {
+func testSubscribe(ctx context.Context, transportFunc func(processor.DocumentTree) error) error {
 	logger := logging.FromContext(ctx)
 	id := uuid.NewV4().String()
 
@@ -208,9 +189,9 @@ func testSubscribe(ctx context.Context, docChannel chan<- processor.DocumentTree
 		doc := processor.Document{}
 		err := json.Unmarshal(d, &doc)
 		if err != nil {
-			fmtErr := fmt.Errorf("[processor: %s] failed unmarshal the document bytes: %v", id, err)
+			fmtErr := fmt.Errorf("[processor: %s] failed unmarshal the document bytes: %w", id, err)
 			logger.Error(fmtErr)
-			return err
+			return fmtErr
 		}
 
 		docNode := &processor.DocumentNode{
@@ -219,7 +200,12 @@ func testSubscribe(ctx context.Context, docChannel chan<- processor.DocumentTree
 		}
 
 		docTree := processor.DocumentTree(docNode)
-		docChannel <- docTree
+		err = transportFunc(docTree)
+		if err != nil {
+			fmtErr := fmt.Errorf("[processor: %s] failed transportFunc: %w", id, err)
+			logger.Error(fmtErr)
+			return fmtErr
+		}
 		logger.Infof("[processor: %s] docTree Processed: %+v", id, docTree.Document.SourceInformation)
 		return nil
 	}
